+++ src/main/java/shop/wannab/frontservice/controller/MypageController.java
@@ -22,22 +22,6 @@ public class MypageController {
     private final UserService userService;
     private final CouponApiClient couponApiClient;
 
-    @GetMapping("/mypage")
-    public String mypageEdit(HttpServletRequest request, Model model) {
-        model.addAttribute("currentUri", request.getRequestURI());
-        UserPageResponse response = userService.readUser();
-        UserViewModel viewModel = UserViewModel.builder()
-                .id(response.username())
-                .password(response.password())
-                .phone(response.phone())
-                .birth(response.birth())
-                .nickname(response.nickname())
-                .email(response.email())
-                .name(response.name())
-                .build();
-        model.addAttribute("user", viewModel);
-        return "user/mypage-edit";
-    }
 
 //    @GetMapping("/mypage-order")
 //    public String mypageOrder(HttpServletRequest request, Model model) {
