+++ src/main/java/shop/wannab/frontservice/user/service/PointService.java
@@ -2,6 +2,8 @@ package shop.wannab.frontservice.user.service;
 
 import jakarta.validation.Valid;
 import java.util.List;
+import shop.wannab.frontservice.category.controller.response.PageResponse;
+import shop.wannab.frontservice.user.dto.PointHistoryResponse;
 import shop.wannab.frontservice.user.dto.PointPageResponse;
 import shop.wannab.frontservice.user.dto.PointPolicyCreateForm;
 import shop.wannab.frontservice.user.dto.PointPolicyUpdateForm;
@@ -12,4 +14,6 @@ public interface PointService {
     void updatePointPolicy(@Valid PointPolicyUpdateForm pointPolicyUpdateForm);
 
     void createPointPolicy(@Valid PointPolicyCreateForm pointPolicyUpdateForm);
+
+    PageResponse<PointHistoryResponse> readPointHistories(int page);
 }
