+++ src/main/resources/templates/user/my-page-point-histories.html
@@ -1,0 +1,54 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org"
+      th:replace="~{layout/user-layout :: layout (
+      '마이 페이지',
+      ~{::body},
+      ~{::head},
+      false
+    )}">
+
+<head>
+    <title>마이 페이지</title>
+    <link rel="stylesheet" th:href="@{/css/user/mypage-address.css}" />
+    <script th:src="@{/js/mypage.js}"></script>
+</head>
+
+<body>
+<div class="mypage-wrapper">
+    <!-- 사이드바 -->
+    <div th:replace="~{components/mypage-sidebar :: mypageSidebar}"></div>
+
+    <!-- 본문 -->
+    <section class="mypage-content">
+        <h3>포인트 내역</h3>
+        <ul class="address-list">
+            <li class="address-item" th:each="history : ${pointHistories.content}">
+                <div class="address-info">
+                    <span class="addressName" th:text="${#temporals.format(history.createdAt, 'yyyy-MM-dd HH:mm')}">날짜</span>
+                    <p class="address" th:text="${history.pointHistoryReason}">사유</p>
+                    <span class="detailAddress" th:text="${history.pointHistoryChange}"
+                          th:classappend="${history.pointHistoryChange > 0} ? 'text-green-600' : 'text-red-500'">+100</span>
+                    <p class="address" th:text="'총 보유: ' + ${history.totalPoints}">총 보유 포인트</p>
+                </div>
+            </li>
+        </ul>
+
+        <!-- 페이지네이션 -->
+        <div class="mt-4 flex justify-center text-sm text-sky-700">
+            <a th:if="${pointHistories.hasPrevious()}"
+               th:href="@{/point-histories(page=${pointHistories.number - 1})}"
+               class="hover:underline">이전</a>
+            <a th:unless="${pointHistories.hasPrevious()}" class="invisible">이전</a>
+
+            <span th:text="${pointHistories.number + 1} + ' / ' + ${pointHistories.totalPages}"
+                  class="inline-block text-center w-24"></span>
+
+            <a th:if="${pointHistories.hasNext()}"
+               th:href="@{/point-histories(page=${pointHistories.number + 1})}"
+               class="hover:underline">다음</a>
+            <a th:unless="${pointHistories.hasNext()}" class="invisible">다음</a>
+        </div>
+    </section>
+</div>
+</body>
+</html>
